+++ frontend/features/template/components/edit/config-component/config-dnd-item.tsx
